Combine tests into less files
