+++ tests/test_gdata.py
@@ -1,13 +1,18 @@
-"""Comprehensive tests for GData class."""
+"""Tests for GData class."""
 
 from __future__ import annotations
 
+import json
 import os
+import sys
+import types
+
 import numpy as np
 import pytest
 
 import postgkyl as pg
 from postgkyl.data.gdata import GData
+import postgkyl.utils.gkeyll_enums as gkenums
 
 
 dir_path = f"{os.path.dirname(__file__)}/test_data"
@@ -260,9 +265,83 @@ class TestGDataInfo:
         info = d.info()
         assert "DG" in info
 
+    def test_info_with_basis_info_modal(self):
+        d = _make([np.linspace(0.0, 1.0, 5)], np.ones((4, 8)))
+        d.ctx.update({
+            "grid_type": "uniform",
+            "lower": np.array([0.0]),
+            "upper": np.array([1.0]),
+            "cells": np.array([4]),
+            "poly_order": 1,
+            "basis_type": "serendipity",
+            "is_modal": True,
+        })
+        info_str = d.info()
+        assert "Basis Type" in info_str
+        assert "modal" in info_str
+
+    def test_info_with_basis_info_nodal(self):
+        d = _make([np.linspace(0.0, 1.0, 5)], np.ones((4, 8)))
+        d.ctx.update({
+            "grid_type": "uniform",
+            "poly_order": 2,
+            "basis_type": "tensor",
+            "is_modal": False,
+        })
+        info_str = d.info()
+        assert "Basis Type" in info_str
+
+    def test_info_with_build_info(self):
+        d = _make([np.linspace(0.0, 1.0, 5)], np.ones((4, 1)))
+        d.ctx.update({
+            "grid_type": "uniform",
+            "changeset": "abc123",
+            "builddate": "2024-01-01",
+        })
+        info_str = d.info()
+        assert "Created with Gkeyll" in info_str
+        assert "abc123" in info_str
+        assert "2024-01-01" in info_str
+
+    def test_info_with_geometry_info(self):
+        d = _make([np.linspace(0.0, 1.0, 5)], np.ones((4, 1)))
+        d.ctx.update({
+            "grid_type": "uniform",
+            "geometry_type": 0,
+            "geqdsk_sign_convention": 1,
+        })
+        info_str = d.info()
+        assert "Geometry info" in info_str
+
+    def test_info_extra_ctx_keys(self):
+        d = _make([np.linspace(0.0, 1.0, 5)], np.ones((4, 1)))
+        d.ctx.update({
+            "grid_type": "uniform",
+            "custom_key": "custom_value",
+        })
+        info_str = d.info()
+        assert "custom_key" in info_str
+
+    def test_info_multicomp(self):
+        d = _make([np.linspace(0.0, 1.0, 5)], np.ones((4, 3)))
+        d.ctx["grid_type"] = "uniform"
+        info_str = d.info()
+        assert "components" in info_str.lower()
+
+    def test_info_with_lower_upper_cells(self):
+        d = _make([np.linspace(0.0, 1.0, 5)], np.ones((4, 1)))
+        d.ctx.update({
+            "grid_type": "uniform",
+            "lower": np.array([0.0]),
+            "upper": np.array([1.0]),
+            "cells": np.array([4]),
+        })
+        info_str = d.info()
+        assert "Lower" in info_str
+
 
 # ---------------------------------------------------------------------------
-# Load from files (using existing test data)
+# Load from files
 # ---------------------------------------------------------------------------
 
 class TestGDataFromFile:
@@ -298,6 +377,85 @@ class TestGDataFromFile:
 
 
 # ---------------------------------------------------------------------------
+# set_neighbors
+# ---------------------------------------------------------------------------
+
+class TestSetNeighbors:
+    def test_set_neighbors_1d_finds_adjacent(self):
+        grid0 = [np.linspace(0.0, 1.0, 5)]
+        grid1 = [np.linspace(1.0, 2.0, 5)]
+        values = np.ones((4, 1))
+        block0 = _make(grid0, values)
+        block1 = _make(grid1, values)
+        block0.set_neighbors([block0, block1])
+        assert block0._neighbors[0][1] is block1
+
+    def test_set_neighbors_1d_finds_left(self):
+        grid0 = [np.linspace(0.0, 1.0, 5)]
+        grid1 = [np.linspace(1.0, 2.0, 5)]
+        values = np.ones((4, 1))
+        block0 = _make(grid0, values)
+        block1 = _make(grid1, values)
+        block1.set_neighbors([block0, block1])
+        assert block1._neighbors[0][0] is block0
+
+    def test_set_neighbors_no_neighbors(self):
+        grid0 = [np.linspace(0.0, 1.0, 5)]
+        values = np.ones((4, 1))
+        block0 = _make(grid0, values)
+        block0.set_neighbors([block0])
+        assert block0._neighbors[0][0] is None
+        assert block0._neighbors[0][1] is None
+
+    def test_set_neighbors_2d(self):
+        grid0 = [np.linspace(0.0, 1.0, 4), np.linspace(0.0, 1.0, 4)]
+        grid1 = [np.linspace(1.0, 2.0, 4), np.linspace(0.0, 1.0, 4)]
+        values = np.ones((3, 3, 1))
+        block0 = _make(grid0, values)
+        block1 = _make(grid1, values)
+        block0.set_neighbors([block0, block1])
+        assert block0._neighbors[0][1] is block1
+
+
+# ---------------------------------------------------------------------------
+# get_num_comps
+# ---------------------------------------------------------------------------
+
+class TestGDataNumComps:
+    def test_num_comps_from_values_after_push(self):
+        d = GData()
+        grid = [np.linspace(0.0, 1.0, 4)]
+        values = np.ones((3, 5))
+        d.push(grid, values)
+        assert d.get_num_comps() == 5
+
+    def test_num_comps_from_ctx_matches_values(self):
+        d = GData()
+        d.ctx["num_comps"] = 5
+        grid = [np.linspace(0.0, 1.0, 4)]
+        values = np.ones((3, 5))
+        d.push(grid, values)
+        assert d.get_num_comps() == 5
+
+    def test_num_comps_direct_values_access(self):
+        d = GData()
+        d._values = np.ones((3, 7))
+        assert d.get_num_comps() == 7
+
+    def test_num_comps_no_values(self):
+        d = GData()
+        assert d.get_num_comps() == 0
+
+    def test_num_comps_ctx_set_to_different_before_push(self):
+        d = GData()
+        d.ctx["num_comps"] = 3
+        grid = [np.linspace(0.0, 1.0, 4)]
+        values = np.ones((3, 5))
+        d.push(grid, values)
+        assert d.get_num_comps() == 5
+
+
+# ---------------------------------------------------------------------------
 # Write
 # ---------------------------------------------------------------------------
 
@@ -326,6 +484,109 @@ class TestGDataWrite:
         out_file = str(tmp_path / "test_write.gkyl")
         d.write(out_name=out_file, extension="gkyl")
         assert os.path.exists(out_file)
-        # Reload and verify
         d2 = pg.GData(out_file)
         np.testing.assert_allclose(d2.get_values(), values)
+
+
+# ---------------------------------------------------------------------------
+# Write VTS series sidecar
+# ---------------------------------------------------------------------------
+
+class _FakeStructuredGrid:
+    def __init__(self, _x, _y, _z):
+        self._point_data = {}
+
+    def __setitem__(self, key, value):
+        self._point_data[key] = value
+
+    def save(self, file_name):
+        with open(file_name, "w", encoding="utf-8") as fh:
+            fh.write("fake-vts")
+
+
+def _write_vts(tmp_path, stem, suffix, *, time=None, frame=None):
+    grid = [np.array([0.0, 1.0, 2.0])]
+    values = np.array([[1.0], [2.0]])
+    data = GData()
+    data.push(grid, values)
+    if time is not None:
+        data.ctx["time"] = time
+    if frame is not None:
+        data.ctx["frame"] = frame
+    out = tmp_path / f"{stem}_{suffix:04d}.vts"
+    data.write(out_name=str(out), extension="vts")
+    return out
+
+
+def test_write_vts_creates_and_updates_series_sidecar(tmp_path, monkeypatch):
+    fake_module = types.SimpleNamespace(StructuredGrid=_FakeStructuredGrid)
+    monkeypatch.setitem(sys.modules, "pyvista", fake_module)
+    first_out = _write_vts(tmp_path, "solution", 1, time=0.25)
+    second_out = _write_vts(tmp_path, "solution", 2, time=0.50)
+    series_file = tmp_path / "solution.vts.series"
+    assert first_out.exists()
+    assert second_out.exists()
+    assert series_file.exists()
+    with open(series_file, "r", encoding="utf-8") as fh:
+        series_data = json.load(fh)
+    assert series_data["file-series-version"] == "1.0"
+    assert series_data["files"] == [
+        {"name": "solution_0001.vts", "time": 0.25},
+        {"name": "solution_0002.vts", "time": 0.5},
+    ]
+
+
+def test_write_vts_series_uses_frame_then_default_time(tmp_path, monkeypatch):
+    fake_module = types.SimpleNamespace(StructuredGrid=_FakeStructuredGrid)
+    monkeypatch.setitem(sys.modules, "pyvista", fake_module)
+    _write_vts(tmp_path, "framecase", 1, frame=7)
+    _write_vts(tmp_path, "framecase", 2)
+    series_file = tmp_path / "framecase.vts.series"
+    with open(series_file, "r", encoding="utf-8") as fh:
+        series_data = json.load(fh)
+    assert series_data["files"] == [
+        {"name": "framecase_0002.vts", "time": 0.0},
+        {"name": "framecase_0001.vts", "time": 7.0},
+    ]
+
+
+def test_write_vts_series_rewrites_existing_entry_without_duplication(tmp_path, monkeypatch):
+    fake_module = types.SimpleNamespace(StructuredGrid=_FakeStructuredGrid)
+    monkeypatch.setitem(sys.modules, "pyvista", fake_module)
+    _write_vts(tmp_path, "resample", 1, time=0.10)
+    _write_vts(tmp_path, "resample", 2, time=0.20)
+    _write_vts(tmp_path, "resample", 2, time=0.40)
+    series_file = tmp_path / "resample.vts.series"
+    with open(series_file, "r", encoding="utf-8") as fh:
+        series_data = json.load(fh)
+    assert series_data["files"] == [
+        {"name": "resample_0001.vts", "time": 0.1},
+        {"name": "resample_0002.vts", "time": 0.4},
+    ]
+
+
+# ---------------------------------------------------------------------------
+# gkeyll_enums
+# ---------------------------------------------------------------------------
+
+class TestGkeyllEnums:
+    def test_enum_idx_to_key(self):
+        result = gkenums.enum_idx_to_key(gkenums.gkyl_geometry_id, 0)
+        assert result == "GKYL_GEOMETRY_NONE"
+
+    def test_enum_idx_to_key_tokamak(self):
+        result = gkenums.enum_idx_to_key(gkenums.gkyl_geometry_id, 1)
+        assert result == "GKYL_GEOMETRY_TOKAMAK"
+
+    def test_enum_key_to_idx(self):
+        result = gkenums.enum_key_to_idx(gkenums.gkyl_geometry_id, "GKYL_GEOMETRY_NONE")
+        assert result == 0
+
+    def test_enum_key_to_idx_mapc2p(self):
+        result = gkenums.enum_key_to_idx(gkenums.gkyl_geometry_id, "GKYL_GEOMETRY_MAPC2P")
+        assert result == 3
+
+    def test_enum_roundtrip(self):
+        idx = 2
+        key = gkenums.enum_idx_to_key(gkenums.gkyl_geometry_id, idx)
+        assert gkenums.enum_key_to_idx(gkenums.gkyl_geometry_id, key) == idx
