+++ tests/test_tools_pressure_diagnostics.py
@@ -1,4 +1,4 @@
-"""Comprehensive tests for tools.pressure_diagnostics."""
+"""Tests for tools.pressure_diagnostics."""
 
 from __future__ import annotations
 
@@ -7,6 +7,16 @@ import pytest
 
 import postgkyl.tools as tools
 from postgkyl.data.gdata import GData
+from postgkyl.tools.pressure_diagnostics import (
+    _get_pb,
+    _get_sf,
+    get_agyro,
+    get_gkyl_10m_agyro,
+    get_gkyl_10m_p_par,
+    get_gkyl_10m_p_perp,
+    get_p_par,
+    get_p_perp,
+)
 
 
 def _make(grid, values):
@@ -19,15 +29,34 @@ _G1D = [np.array([0.0, 1.0])]
 
 
 def _make_diagonal_pressure(pxx, pyy, pzz):
-    """6-component pressure tensor (no off-diagonal) as tuple."""
     v = np.array([[pxx, 0.0, 0.0, pyy, 0.0, pzz]])
     return _G1D, v
 
 
 def _make_b(bx, by, bz):
-    """3-component B-field as tuple."""
     v = np.array([[bx, by, bz]])
     return _G1D, v
+
+
+def _make_pij_gdata(pxx=1.0, pxy=0.0, pxz=0.0, pyy=1.0, pyz=0.0, pzz=1.0):
+    values = np.array([[pxx, pxy, pxz, pyy, pyz, pzz]])
+    return _make(_G1D, values)
+
+
+def _make_b_gdata(bx=0.0, by=0.0, bz=1.0):
+    values = np.array([[bx, by, bz]])
+    return _make(_G1D, values)
+
+
+def _make_10mom_gdata(rho=1.0, vx=0.0, p_par=1.0, p_perp=0.5):
+    Pxx = p_perp + rho * vx**2
+    values = np.array([[rho, rho * vx, 0.0, 0.0, Pxx, 0.0, 0.0, p_perp, 0.0, p_par]])
+    return _make(_G1D, values)
+
+
+def _make_field_gdata(bx=0.0, by=0.0, bz=1.0):
+    values = np.array([[0.0, 0.0, 0.0, bx, by, bz]])
+    return _make(_G1D, values)
 
 
 # ---------------------------------------------------------------------------
@@ -54,7 +83,6 @@ class TestGetPPar:
         np.testing.assert_allclose(p_par.flat[0], 3.0, rtol=1e-12)
 
     def test_isotropic_pressure_p_par_equals_p(self):
-        # For isotropic p, p_par = p regardless of B direction
         p_val = 2.0
         p_in = _make_diagonal_pressure(p_val, p_val, p_val)
         b_in = _make_b(1.0, 1.0, 0.0)
@@ -62,11 +90,9 @@ class TestGetPPar:
         np.testing.assert_allclose(p_par.flat[0], p_val, rtol=1e-10)
 
     def test_b_diagonal_gives_average(self):
-        # B at 45° in xy, diagonal p
         p_in = _make_diagonal_pressure(1.0, 2.0, 0.0)
         b_in = _make_b(1.0 / np.sqrt(2), 1.0 / np.sqrt(2), 0.0)
         _, p_par = tools.get_p_par(p_in, b_in)
-        # p_par = (bx^2*pxx + by^2*pyy) / |B|^2 = 0.5*1 + 0.5*2 = 1.5
         np.testing.assert_allclose(p_par.flat[0], 1.5, rtol=1e-12)
 
 
@@ -80,7 +106,6 @@ class TestGetPPerp:
         b_in = _make_b(1.0, 0.0, 0.0)
         _, p_par = tools.get_p_par(p_in, b_in)
         _, p_perp = tools.get_p_perp(p_in, b_in)
-        # p_perp = (pxx + pyy + pzz - p_par) / 2 = (1+0.6+0.4 - 1) / 2 = 0.5
         np.testing.assert_allclose(p_perp.flat[0], 0.5, rtol=1e-12)
 
     def test_isotropic_pressure_perp_equals_par(self):
@@ -90,6 +115,13 @@ class TestGetPPerp:
         _, p_par = tools.get_p_par(p_in, b_in)
         _, p_perp = tools.get_p_perp(p_in, b_in)
         np.testing.assert_allclose(p_perp.flat[0], p_val, rtol=1e-10)
+
+    def test_isotropic_via_gdata(self):
+        p = _make_pij_gdata(pxx=1.0, pyy=1.0, pzz=1.0)
+        b = _make_b_gdata(bz=1.0)
+        _, p_par_val = get_p_par(p, b)
+        _, p_perp_val = get_p_perp(p, b)
+        np.testing.assert_allclose(p_par_val.flat[0], p_perp_val.flat[0], atol=1e-10)
 
 
 # ---------------------------------------------------------------------------
@@ -131,8 +163,13 @@ class TestGetAgyro:
         with pytest.raises(ValueError, match="swisdak.*frobenius"):
             tools.get_agyro(p_in, b_in, measure="invalid")
 
+    def test_invalid_measure_raises_via_gdata(self):
+        p = _make_pij_gdata(pxx=2.0, pyy=1.0, pzz=1.0, pxy=0.5)
+        b = _make_b_gdata(bz=1.0)
+        with pytest.raises(ValueError, match="needs to be either"):
+            get_agyro(p, b, measure="invalid")
+
     def test_agyrotropic_swisdak_nonzero(self):
-        # Non-gyrotropic: off-diagonal pxy ≠ 0 breaks gyrotropy
         v = np.array([[2.0, 0.5, 0.0, 1.0, 0.0, 1.0]])
         p_in = (_G1D, v)
         b_in = _make_b(1.0, 0.0, 0.0)
@@ -140,7 +177,6 @@ class TestGetAgyro:
         assert Q.flat[0] > 0.0
 
     def test_agyrotropic_frobenius_nonzero(self):
-        # Non-gyrotropic: off-diagonal pxy ≠ 0
         v = np.array([[2.0, 0.5, 0.0, 1.0, 0.0, 1.0]])
         p_in = (_G1D, v)
         b_in = _make_b(1.0, 0.0, 0.0)
@@ -149,22 +185,64 @@ class TestGetAgyro:
 
 
 # ---------------------------------------------------------------------------
-# get_gkyl_10m_p_par / get_gkyl_10m_p_perp / get_gkyl_10m_agyro
-# (wrappers that take full 10-moment + field data)
+# _get_pb private helper
+# ---------------------------------------------------------------------------
+
+class TestGetPb:
+    def test_returns_9_components(self):
+        p = _make_pij_gdata()
+        b = _make_b_gdata(bz=1.0)
+        result = _get_pb(p, b)
+        assert len(result) == 9
+
+    def test_values_correct(self):
+        p = _make_pij_gdata(pxx=2.0, pxy=0.5, pxz=0.1, pyy=3.0, pyz=0.2, pzz=4.0)
+        b = _make_b_gdata(bx=1.0, by=2.0, bz=3.0)
+        pxx, pxy, pxz, pyy, pyz, pzz, bx, by, bz = _get_pb(p, b)
+        np.testing.assert_allclose(pxx.flat[0], 2.0)
+        np.testing.assert_allclose(pxy.flat[0], 0.5)
+        np.testing.assert_allclose(bx.flat[0], 1.0)
+        np.testing.assert_allclose(bz.flat[0], 3.0)
+
+    def test_with_tuples(self):
+        p_values = np.array([[1.0, 0.5, 0.0, 1.0, 0.0, 1.0]])
+        b_values = np.array([[0.0, 0.0, 1.0]])
+        result = _get_pb((_G1D, p_values), (_G1D, b_values))
+        assert len(result) == 9
+
+
+# ---------------------------------------------------------------------------
+# _get_sf private helper
+# ---------------------------------------------------------------------------
+
+class TestGetSf:
+    def test_returns_4_items(self):
+        species = _make_10mom_gdata()
+        field = _make_field_gdata(bz=1.0)
+        result = _get_sf(species, field)
+        assert len(result) == 4
+
+    def test_b_values_from_field(self):
+        field = _make_field_gdata(bx=3.0, by=4.0, bz=0.0)
+        species = _make_10mom_gdata()
+        p_grid, p_values, b_grid, b_values = _get_sf(species, field)
+        np.testing.assert_allclose(b_values.flat[0], 3.0)
+        np.testing.assert_allclose(b_values.flat[1], 4.0)
+        np.testing.assert_allclose(b_values.flat[2], 0.0)
+
+
+# ---------------------------------------------------------------------------
+# get_gkyl_10m wrappers
 # ---------------------------------------------------------------------------
 
 class TestGkyl10mWrappers:
-    """These wrapper functions unpack pij from the 10-moment array and B from field."""
-
     @staticmethod
     def _make_10m_and_field():
-        # rho=1, vx=0.5, vy=0, vz=0; add pxy_thermal=0.3 to break gyrotropy
         rho, vx = 1.0, 0.5
         Pxx = 2.0 + rho * vx**2
-        Pxy = 0.3 + rho * vx * 0.0  # off-diagonal breaks gyrotropy
+        Pxy = 0.3 + rho * vx * 0.0
         mom10 = np.array([[rho, rho * vx, 0.0, 0.0,
                            Pxx, Pxy, 0.0, 1.0, 0.0, 1.0]])
-        # EM field: [Ex,Ey,Ez,Bx,By,Bz] - B along x
         field_vals = np.array([[0.0, 0.0, 0.0, 1.0, 0.0, 0.0]])
         g = [np.array([0.0, 1.0])]
         species = GData()
@@ -176,22 +254,37 @@ class TestGkyl10mWrappers:
     def test_p_par_wrapper(self):
         species, field = self._make_10m_and_field()
         _, p_par = tools.get_gkyl_10m_p_par(species, field)
-        # pxx_thermal = 2.0, B along x → p_par = pxx_thermal
         np.testing.assert_allclose(p_par.flat[0], 2.0, rtol=1e-10)
 
     def test_p_perp_wrapper(self):
         species, field = self._make_10m_and_field()
         _, p_perp = tools.get_gkyl_10m_p_perp(species, field)
-        # pyy_thermal = pzz_thermal = 1.0 → p_perp = (pyy+pzz-p_par)/2 = (1+1)/2 = 1
         np.testing.assert_allclose(p_perp.flat[0], 1.0, rtol=1e-10)
 
     def test_agyro_wrapper_swisdak(self):
         species, field = self._make_10m_and_field()
         _, Q = tools.get_gkyl_10m_agyro(species, field, measure="swisdak")
-        # anisotropic → Q > 0
         assert Q.flat[0] > 0.0
 
     def test_agyro_wrapper_frobenius(self):
         species, field = self._make_10m_and_field()
         _, Q = tools.get_gkyl_10m_agyro(species, field, measure="frobenius")
         assert Q.flat[0] > 0.0
+
+    def test_get_gkyl_10m_p_par_via_direct_import(self):
+        species = _make_10mom_gdata(p_par=2.0, p_perp=1.0)
+        field = _make_field_gdata(bz=1.0)
+        grid, p_par = get_gkyl_10m_p_par(species, field)
+        assert p_par is not None
+
+    def test_get_gkyl_10m_p_perp_via_direct_import(self):
+        species = _make_10mom_gdata(p_par=2.0, p_perp=1.0)
+        field = _make_field_gdata(bz=1.0)
+        grid, p_perp = get_gkyl_10m_p_perp(species, field)
+        assert p_perp is not None
+
+    def test_get_gkyl_10m_agyro_frobenius_via_direct_import(self):
+        species = _make_10mom_gdata()
+        field = _make_field_gdata(bz=1.0)
+        grid, agyro = get_gkyl_10m_agyro(species, field, measure="frobenius")
+        assert agyro is not None
