+++ tests/test_utils.py
@@ -1,17 +1,92 @@
-"""Tests for utils: gk_utils, verb_print, load_style, gkeyll_enums."""
+"""Tests for postgkyl utilities."""
 
 from __future__ import annotations
 
 import os
+import time
+
+import click
 import numpy as np
 import pytest
-import click
 
-from postgkyl.utils.gk_utils import parse_slice_string, get_block_indices
-from postgkyl.utils.gk_utils import read_gfile
+import postgkyl as pg
+from postgkyl.data.gdata import GData
+from postgkyl.utils.gk_utils import get_block_indices, parse_slice_string, read_gfile
+from postgkyl.utils.input_parser import input_parser
 
 
 dir_path = f"{os.path.dirname(__file__)}/test_data"
+
+
+# ---------------------------------------------------------------------------
+# input_parser
+# ---------------------------------------------------------------------------
+
+class TestInputParser:
+    def test_gdata_returns_grid_and_values(self):
+        d = GData()
+        grid = [np.linspace(0.0, 1.0, 4)]
+        values = np.ones((3, 1))
+        d.push(grid, values)
+        g, v = input_parser(d)
+        assert g is d.get_grid()
+        assert v is d.get_values()
+
+    def test_numpy_array_returns_empty_grid(self):
+        arr = np.array([1.0, 2.0, 3.0])
+        g, v = input_parser(arr)
+        assert g == ()
+        assert v is arr
+
+    def test_tuple_of_grid_and_values(self):
+        grid = [np.array([0.0, 1.0])]
+        values = np.array([[1.0]])
+        g, v = input_parser((grid, values))
+        assert g is grid
+        assert v is values
+
+    def test_list_of_grid_and_values(self):
+        grid = [np.array([0.0, 1.0])]
+        values = np.array([[1.0]])
+        g, v = input_parser([grid, values])
+        assert g is grid
+        assert v is values
+
+    def test_tuple_grid_must_be_list_raises(self):
+        with pytest.raises(TypeError, match="grid"):
+            input_parser((np.array([0.0, 1.0]), np.array([[1.0]])))
+
+    def test_tuple_values_must_be_ndarray_raises(self):
+        grid = [np.array([0.0, 1.0])]
+        with pytest.raises(TypeError, match="values"):
+            input_parser((grid, [[1.0]]))
+
+    def test_tuple_wrong_length_raises(self):
+        with pytest.raises(TypeError):
+            input_parser(([np.array([0.0])], np.array([[1.0]]), "extra"))
+
+    def test_wrong_type_raises(self):
+        with pytest.raises(TypeError):
+            input_parser("a_string")
+
+    def test_integer_raises(self):
+        with pytest.raises(TypeError):
+            input_parser(42)
+
+    def test_2d_grid_values_tuple(self):
+        grid = [np.linspace(0.0, 1.0, 4), np.linspace(0.0, 1.0, 3)]
+        values = np.ones((3, 2, 1))
+        g, v = input_parser((grid, values))
+        assert len(g) == 2
+        assert v.shape == (3, 2, 1)
+
+    def test_dim_mismatch_raises(self):
+        grid = [np.linspace(0.0, 1.0, 4),
+                np.linspace(0.0, 1.0, 3),
+                np.linspace(0.0, 1.0, 3)]
+        values = np.ones((5, 1))
+        with pytest.raises(ValueError):
+            input_parser((grid, values))
 
 
 # ---------------------------------------------------------------------------
@@ -57,7 +132,6 @@ class TestGetBlockIndices:
         assert blocks == [0]
 
     def test_all_blocks_no_files(self, tmp_path):
-        # No files match → 0 blocks
         pattern = str(tmp_path / "*.gkyl")
         blocks = get_block_indices("-1", pattern)
         assert blocks == []
@@ -85,7 +159,6 @@ class TestGetBlockIndices:
 
 class TestReadGfile:
     def test_read_gfile_dynvector(self):
-        # twostream-field-energy.gkyl is a dynvector (simple 1D file)
         fn = f"{dir_path}/twostream-field-energy.gkyl"
         if not os.path.exists(fn):
             pytest.skip("test data not available")
@@ -106,7 +179,6 @@ class TestReadGfile:
 
 class TestVerbPrint:
     def test_verb_print_verbose_true(self, capsys):
-        import time
         from postgkyl.utils.verb_print import verb_print
 
         ctx = click.core.Context(click.Command("test"))
@@ -115,11 +187,8 @@ class TestVerbPrint:
             "start_time": time.time(),
         }
         verb_print(ctx, "test message")
-        # click.echo writes to stdout; capsys may or may not capture it
-        # but the function should not raise
 
     def test_verb_print_verbose_false(self):
-        import time
         from postgkyl.utils.verb_print import verb_print
 
         ctx = click.core.Context(click.Command("test"))
@@ -127,7 +196,6 @@ class TestVerbPrint:
             "verbose": False,
             "start_time": time.time(),
         }
-        # Should not raise, and should not print anything
         verb_print(ctx, "test message")
 
 
@@ -138,11 +206,9 @@ class TestVerbPrint:
 class TestLoadStyle:
     def test_load_style_simple_key(self, tmp_path):
         from postgkyl.utils.load_style import load_style
-        import click
 
         style_file = tmp_path / "style.rc"
         style_file.write_text("lines.linewidth: 2\n")
-
         ctx = click.core.Context(click.Command("test"))
         ctx.obj = {"rcParams": {}}
         load_style(ctx, str(style_file))
@@ -151,11 +217,9 @@ class TestLoadStyle:
 
     def test_load_style_multiple_keys(self, tmp_path):
         from postgkyl.utils.load_style import load_style
-        import click
 
         style_file = tmp_path / "style.rc"
         style_file.write_text("lines.linewidth: 2\nfont.size: 12\n")
-
         ctx = click.core.Context(click.Command("test"))
         ctx.obj = {"rcParams": {}}
         load_style(ctx, str(style_file))
